+++ src/Symfony/Component/Form/Extension/Csrf/CsrfExtension.php
@@ -13,7 +13,6 @@ namespace Symfony\Component\Form\Extension\Csrf;
 
 use Symfony\Component\Form\AbstractExtension;
 use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
-use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
@@ -27,16 +26,8 @@ class CsrfExtension extends AbstractExtension
     private $translator;
     private $translationDomain;
 
-    /**
-     * @param CsrfTokenManagerInterface $tokenManager      The CSRF token manager
-     * @param TranslatorInterface|null  $translator        The translator for translating error messages
-     * @param string|null               $translationDomain The translation domain for translating
-     */
-    public function __construct(CsrfTokenManagerInterface $tokenManager, $translator = null, string $translationDomain = null)
+    public function __construct(CsrfTokenManagerInterface $tokenManager, TranslatorInterface $translator = null, string $translationDomain = null)
     {
-        if (null !== $translator && !$translator instanceof LegacyTranslatorInterface && !$translator instanceof TranslatorInterface) {
-            throw new \TypeError(sprintf('Argument 2 passed to %s() must be an instance of %s, %s given.', __METHOD__, TranslatorInterface::class, \is_object($translator) ? \get_class($translator) : \gettype($translator)));
-        }
         $this->tokenManager = $tokenManager;
         $this->translator = $translator;
         $this->translationDomain = $translationDomain;