+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/FileTypeTest.php
@@ -17,7 +17,7 @@ use Symfony\Component\Form\NativeRequestHandler;
 use Symfony\Component\Form\RequestHandlerInterface;
 use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
-use Symfony\Component\Translation\TranslatorInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class FileTypeTest extends BaseTypeTest
 {