+++ src/Symfony/Component/Translation/IdentityTranslator.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Translation;
 
-use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;
+use Symfony\Contracts\Translation\LocaleAwareInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Symfony\Contracts\Translation\TranslatorTrait;
 
@@ -20,42 +20,7 @@ use Symfony\Contracts\Translation\TranslatorTrait;
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-class IdentityTranslator implements LegacyTranslatorInterface, TranslatorInterface
+class IdentityTranslator implements TranslatorInterface, LocaleAwareInterface
 {
     use TranslatorTrait;
-
-    private $selector;
-
-    /**
-     * @param MessageSelector|null $selector The message selector for pluralization
-     */
-    public function __construct(MessageSelector $selector = null)
-    {
-        $this->selector = $selector;
-
-        if (__CLASS__ !== \get_class($this)) {
-            @trigger_error(sprintf('Calling "%s()" is deprecated since Symfony 4.2.', __METHOD__), E_USER_DEPRECATED);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @deprecated since Symfony 4.2, use the trans() method instead with a %count% parameter
-     */
-    public function transChoice($id, $number, array $parameters = [], $domain = null, $locale = null)
-    {
-        @trigger_error(sprintf('The "%s()" method is deprecated since Symfony 4.2, use the trans() one instead with a "%%count%%" parameter.', __METHOD__), E_USER_DEPRECATED);
-
-        if ($this->selector) {
-            return strtr($this->selector->choose((string) $id, $number, $locale ?: $this->getLocale()), $parameters);
-        }
-
-        return $this->trans($id, ['%count%' => $number] + $parameters, $domain, $locale);
-    }
-
-    private function getPluralizationRule(int $number, string $locale): int
-    {
-        return PluralizationRules::get($number, $locale, false);
-    }
 }