+++ src/Symfony/Component/Translation/Tests/Extractor/PhpExtractorTest.php
@@ -50,7 +50,6 @@ EOF;
                 'single-quoted key with "quote mark at the end"' => 'prefixsingle-quoted key with "quote mark at the end"',
                 $expectedHeredoc => 'prefix'.$expectedHeredoc,
                 $expectedNowdoc => 'prefix'.$expectedNowdoc,
-                '{0} There is no apples|{1} There is one apple|]1,Inf[ There are %count% apples' => 'prefix{0} There is no apples|{1} There is one apple|]1,Inf[ There are %count% apples',
                 'concatenated message with heredoc and nowdoc' => 'prefixconcatenated message with heredoc and nowdoc',
                 'default domain' => 'prefixdefault domain',
             ],
@@ -59,11 +58,7 @@ EOF;
                 'other-domain-test-no-params-long-array' => 'prefixother-domain-test-no-params-long-array',
                 'other-domain-test-params-short-array' => 'prefixother-domain-test-params-short-array',
                 'other-domain-test-params-long-array' => 'prefixother-domain-test-params-long-array',
-                'other-domain-test-trans-choice-short-array-%count%' => 'prefixother-domain-test-trans-choice-short-array-%count%',
-                'other-domain-test-trans-choice-long-array-%count%' => 'prefixother-domain-test-trans-choice-long-array-%count%',
                 'typecast' => 'prefixtypecast',
-                'msg1' => 'prefixmsg1',
-                'msg2' => 'prefixmsg2',
             ],
         ];
         $actualCatalogue = $catalogue->all();
@@ -72,7 +67,7 @@ EOF;
 
         $filename = str_replace(\DIRECTORY_SEPARATOR, '/', __DIR__).'/../fixtures/extractor/translation.html.php';
         $this->assertEquals(['sources' => [$filename.':2']], $catalogue->getMetadata('single-quoted key'));
-        $this->assertEquals(['sources' => [$filename.':43']], $catalogue->getMetadata('other-domain-test-no-params-short-array', 'not_messages'));
+        $this->assertEquals(['sources' => [$filename.':37']], $catalogue->getMetadata('other-domain-test-no-params-short-array', 'not_messages'));
     }
 
     public function resourcesProvider()