+++ src/Symfony/Component/Translation/Writer/TranslationWriter.php
@@ -37,22 +37,6 @@ class TranslationWriter implements TranslationWriterInterface
     }
 
     /**
-     * Disables dumper backup.
-     *
-     * @deprecated since Symfony 4.1
-     */
-    public function disableBackup()
-    {
-        @trigger_error(sprintf('The "%s()" method is deprecated since Symfony 4.1.', __METHOD__), E_USER_DEPRECATED);
-
-        foreach ($this->dumpers as $dumper) {
-            if (method_exists($dumper, 'setBackup')) {
-                $dumper->setBackup(false);
-            }
-        }
-    }
-
-    /**
      * Obtains the list of supported formats.
      *
      * @return array